requirements.txt
dimod>=0.8.11; python_version!="3.4"
numpy>=1.15.4; python_version!="3.4"

minorminer>=0.1.7
dwave-system>=0.7.6
dwave-neal>=0.5.1
dwave-tabu>=0.2.0
networkx
dwave-networkx>=0.8.6
click
plucky>=0.4.3
